ImageWriter

Description

Documentation

type Color f = [f]

A list to store components of a color (e.g. r,g,b values)

write_to_ppm

Arguments

:: (Floating f, RealFrac f)

=> Int
The maximum pixel color value

-> Image f
The image to print

-> String
The ppm string

Returns a string with the contents of a ppm file for the given image

data Image f

An image specified by a width and height (in pixels) and a sequence of pixel color values

Constructors

Image Int Int (Seq (Color f))