Interval

Description

Documentation

data Interval a

An interval (range between numbers)

Constructors

Interval a a

to_tuple :: Interval a -> (a, a)

Convert an interval into a tuple containing (min, max)

from_tuple :: (a, a) -> Interval a

Create a tuple (min, max) into the interval [min, max]

from_num :: a -> Interval a

Convert a value into an interval that only includes itself

midpoint :: (Num a, Fractional a) => Interval a -> a

Compute the midpoint of an interval (max - min) / 2

corner :: (Num a, Ord a) => Interval a -> a

Computes the corner of an interval. If the interval includes 0, the corner is 0. Otherwise, it is the bound of the interval that is closest to 0.

width :: Num a => Interval a -> a

Computes the width of an interval (max - min)

radius :: (Num a, Fractional a) => Interval a -> a

Computes the radius of an interval (half the width)

abs :: (Num a, Ord a) => Interval a -> Interval a

Computes the interval which has the absolute values of the bounds of the input interval as bounds

mig :: (Num a, Ord a) => Interval a -> a

Computes the mignitude of an interval (the minimum element of the absolute value of the interval)

mag :: (Num a, Ord a) => Interval a -> a

Computes the magnitude of an interval (the maximum element of the absolute value of the interval)

sign :: (Num a, Ord a) => Interval a -> a

Computes the sign of an interval (0 if the interval includes 0, -1 if the interval is all below 0, and 1 if the interval is all above 0)

hull :: (Num a, Ord a) => [a] -> Interval a

Generates the interval hull of a list of inputs (an interval that contains them all)

neg :: Num a => Interval a -> Interval a

Generates an interval whose elements are the negations of the elements of the input interval

plus :: Num a => Interval a -> Interval a -> Interval a

Generates an interval whose elements are the sums of elements of the input intervals

minus :: Num a => Interval a -> Interval a -> Interval a

Generates an interval whose elements are the differences of elements in the input intervals (the elements from the second interval are subttracted from those from the first interval).

times :: (Num a, Ord a) => Interval a -> Interval a -> Interval a

Generates an interval whose elements are the products of elements in the input intervals

scalar_mult :: (Num a, Ord a) => a -> Interval a -> Interval a

Generates an interval whose elements are the elements of the input interval multiplied by some scalar

divided :: (Num a, Ord a, Fractional a) => Interval a -> Interval a -> Interval a

Generates an interval whose elements are the quotients of elements in the input intervals (the elements of the first interval are in the numerator and the elements of the second are in the denominator)

pow :: (Num a, Ord a, Fractional a) => Interval a -> Int -> Interval a

Generates the interval whose elements are the powers of elements in the input interval

contains :: (Num a, Ord a) => Interval a -> a -> Bool

Returns whether or not an interval contains a given value