IntervalNewton

Description

Documentation

inclusion

Arguments

:: (Eq a, Ord a, Num a)

=> Polynomial a
The polynomial whose inclusion function we want to find

-> [Char]
The variables in the polynomial

-> [Interval a]
The intervals over which the inclusion is to be calculated

-> Interval a
The interval given by the inclusion function

Computes the inclusion function of a polynomial applied to a given interval by computing the polynomial's corner taylor form. If any variables of the polynomial are not specified, they are assumed to be 0.

zeros

Arguments

:: (Num a, Ord a, Fractional a, Show a)

=> Polynomial a
The polynomial whose roots we are looking for

-> [Char]
The variables in the polynomial

-> [Interval a]
The bounds to search in

-> a
The maximum box-width to returj

-> [[Interval a]]
The boxes where the roots might be

Runs Interval Newton to find regions where the polynomials's roots are

leaves :: (Num a, Ord a, Fractional a, Show a) => Polynomial a -> [Char] -> [Interval a] -> a -> [KDTree a (Interval a)]

Runs Interval Newton to find regions where the polynomials's roots are Returns a list of the leaves that we generate (which contain the boundaries of the regions they represent and the inclusion function evaluated on those regions.

write_to_image

Arguments

:: (Num a, Fractional a, RealFrac a, Num b, Show a)

=> [[Interval a]]
The intervals to draw

-> a
The width of the region the intervals are defined on

-> a
The height of the region the intervals are defined on

-> a
The x offset of the region the intervals are defined on

-> a
The y offset of the region the intervals are defined on

-> a
The desired image width

-> a
The desired image height

-> Image b

Returns an Image that draws the given intervals. Only draws the intervals using the first two coordinates if more are present

write_leaf_data :: (Num a, Fractional a, RealFrac a, Show a) => [KDTree a (Interval a)] -> String

Returns a string containing the information from the leaves. Leaves are separated by newlines. Each line contains the low value of the leaf's interval, the high value of the leaf's interval, and the low and high values of the boundaries in each dimension. All values are separated by commas.