KDTree

Description

Documentation

data KDTree a b

A KD Tree that stores data in its leaves

construct_simple_KDTree :: [(a, a)] -> b -> KDTree a b

Constructs a KDTree consisting of a single leaf with the given boundaries and other value

construct_KDTree :: KDTree a b -> KDTree a b -> KDTree a b

Construct a KDTree with 2 children

split :: (Num a, Ord a, Fractional a) => KDTree a b -> (KDTree a b, KDTree a b)

Splits a leaf into two leaves. If passed a non-leaf, returns 2 copies of the non-leaf

is_leaf :: KDTree a b -> Bool

Checks whether a given KDTree is a leaf node

width :: (Num a, Ord a) => KDTree a b -> a

If passed a leaf node, returns the node's width. Otherwise, returns -1

set_leaf_val :: KDTree a b -> b -> KDTree a b

If passed a leaf, sets the leaf's value. Otherwise does nothing

get_leaf_val :: KDTree a b -> b

If passed a leaf, returns the leaf's value. Otherwise, returns the first child's value

leaf_boundaries :: KDTree a b -> [(a, a)]

If passed a leaf, returns the a list of the leaf's boundaries. Otherwise, returns an empty list.

children :: KDTree a b -> [KDTree a b]

Returns a list of a node's children