Polynomial

Description

Documentation

data Polynomial a

A polynomial

Constructors

Polynomial [Term a]
A plain polynomial

Product (Polynomial a) (Polynomial a)
A polynomial expressed as a product of polynomials

Power (Polynomial a) Int
A polynomial expressed as a power of a polynomial

Sum (Polynomial a) (Polynomial a)
A polynomial expressed as a sum of polynomials

data Term a

A term in a polynomial

Constructors

Term a [Atom]

data Atom

A variable raised to a power. Used to form polynomial Terms

Constructors

Atom Char Int

to_const :: (Num a, Ord a) => Polynomial a -> a

Returns the constant term of the polynomial

const_poly :: a -> Polynomial a

Construct a constant polynomial

x_poly :: Num a => [a] -> Polynomial a

Construct a polynomial in x from a list of coefficients

construct_poly :: (Num a, Read a, Ord a) => String -> Polynomial a

Construct a polynomial from a string. The parsing is very basic so every | term needs a coefficient (even if it's 1) and a power of x (even if it's x^0 | or x^1)

construct_univariate_polynomial :: Num a => Char -> [a] -> Polynomial a

Construct a polynomial in one arbitrary variable from a list of coefficients

replace_variable :: (Eq a, Num a, Ord a) => Char -> Polynomial a -> Polynomial a -> Polynomial a

Replace a variable in a polynomial with another polynomial

evaluate :: (Eq a, Num a, Ord a) => [Char] -> [a] -> Polynomial a -> Polynomial a

Evaluate a polynomial with given values for some of its variables

evaluate_one_var :: (Eq a, Num a, Ord a) => Char -> a -> Polynomial a -> Polynomial a

Evaluate a polynomial at var = val

variables_in_poly :: Polynomial a -> [Char]

Return a list of the variables used in a polynomial

taylor_expand :: (Num a, Ord a) => [Char] -> [a] -> Polynomial a -> Polynomial a

If we write x as (c + (x-c)), we can find the taylor expansion of p(x) by writing | p(c + (x-c)) and expanding. We use this technique to find the taylor expansion of | a polynomial

partials :: (Num a, Ord a) => [Char] -> Polynomial a -> [Polynomial a]

Return the partial derivatives with return to the given variables

differentiate :: (Num a, Ord a) => Char -> Polynomial a -> Polynomial a

Differentiate a polynomial with respect to a variable

hessian_product :: Num a => Polynomial a -> a -> a

Computes (x-c) H(p(x)) (x-c)^T where x is a vector and H(p(x)) is the Hessian matrix of polynomial p(x). (Not yet implemented)

add :: (Num a, Ord a) => Polynomial a -> Polynomial a -> Polynomial a

Return the formal sum of two polynomials

scalar_multiply :: (Num a, Ord a) => a -> Polynomial a -> Polynomial a

Multiply a polynomial by a constant

multiply :: Eq a => Polynomial a -> Polynomial a -> Polynomial a

Return the formal product of two polynomials

reduce_full :: (Num a, Ord a, Eq a) => Polynomial a -> Polynomial a

Return the expanded form of the polynomial

add_full :: (Num a, Ord a) => Polynomial a -> Polynomial a -> Polynomial a

Return the expanded sum of polynomials

multiply_full :: (Num a, Ord a, Eq a) => Polynomial a -> Polynomial a -> Polynomial a

Return the expanded product of polynomials

reduce_sums :: (Num a, Ord a, Eq a) => Polynomial a -> Polynomial a

Sums together terms with the same variables. Does not multiply out products