PolynomialParser

Description

Documentation

parse_polynomial :: (Num a, Read a) => String -> Maybe (Polynomial a)

Reads in a string and returns a Polynomial if it parsed correctly and Nothing if the parser failed